requirements.txt
# Core dependencies
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Backend API dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Databricks SDK and connectors
databricks-sql-connector>=3.0.0
databricks-sdk>=0.33.0

# PDF and document parsing (Lamp)
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# LLM and API clients
openai>=1.0.0
aiohttp>=3.9.0

# Data processing
pandas>=2.1.0

# Utilities
sqlparse>=0.4.0
tqdm>=4.66.0
PyJWT>=2.8.0

# Development
pytest>=7.4.0
